+++ src/HighTable.tsx
@@ -4,7 +4,7 @@ import Row from './components/Row.js'
 import RowHeader from './components/RowHeader.js'
 import { DataFrame } from './dataframe.js'
 import { rowError } from './helpers/text.js'
-import { useInputState } from './hooks.js'
+import { useInputState } from './hooks/useInputState.js'
 import { PartialRow } from './row.js'
 import { Selection, SortIndex, areAllSelected, computeNewSelection, isSelected, toggleAll } from './selection.js'
 import { OrderBy, areEqualOrderBy } from './sort.js'
